pyproject.toml
[project]
name = "snipster"
version = "0.1.0"
description = "Intermediate Cohort Project"
readme = "README.md"
authors = [
    { name = "jjeg1979", email = "jjeg1979@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pytest-mock>=3.14.1",
    "python-decouple>=3.8",
    "ruff>=0.11.13",
    "sqlmodel>=0.0.24",
]

[project.scripts]
snipster = "snipster:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "poethepoet>=0.35.0",
    "pytest-cov>=6.2.1",
    "ty>=0.0.1a11",
]


[tool.poe.tasks]
test         = "pytest -vv --cov=src --cov-report=term-missing --cov-report=html"
lint         = "ruff check src"
format       = "ruff format src"
format-check = "ruff check src --fix"
type-check   = "ruff typecheck src"
run          = "snipster"
tc           = "ty check ."
